Change routing server for use in Singularity env
Changes OsrmEngine.py to be compatible with use in an environment
where the osrm-backend is contained in a Singularity image rather
than executables.

These settings are governed in local.py, which now contains the
variables use_singularity and simg_loc. If using on an existing
platform, simply set use_singularity to False and there should be
no issue.